silence a couple more warnings
git-svn-id: https://svn.code.sf.net/p/ec2drv/code/ec2drv/trunk@193 07914c0c-2d3e-0410-adf3-90c1a7ff5250

+++ src/ec2tools/ec2test-any.cpp
@@ -865,10 +865,10 @@ bool test_debug( EC2DRV &obj )
 		0x00,				// 0x0008	NOP				BP_2
 		0x00,				// 0x0009	NOP
 		0x00,				// 0x000a	NOP
-		0xD8,-10,			// 0x000b	DJNZ R0, loop1
+		0xD8,uint8_t(-10),	// 0x000b	DJNZ R0, loop1
 		0x02, 0x00, 0x00	// 0x000d	LJMP start (0x0000)
 	};
-	
+
 	const uint32_t BP_0 = 0x0005;
 	const uint32_t BP_1 = 0x0007;
 	const uint32_t BP_2 = 0x0008;
